Adding changes required for this code to be used in IDC Inflow

+++ base/3_gpu_drivers_setup.sh
@@ -71,6 +71,9 @@ sudo apt-get install -y \
   level-zero-dev
 
 # Reboot
-colored_output "Rebooting the system in 10 seconds..." blue
-sleep 10
-sudo reboot
+if [ -z "$OMMIT_REBOOT" ]
+then
+    colored_output "Rebooting the system in 10 seconds..." blue
+    sleep 10
+    sudo reboot
+fi
